Fun7 para sortear uma marmita, usando a base de marmitas

Fun7 para sortear uma marmita, usando a base de marmitas

sortear_marmita(tipo = c("Sopa", "Prato", "Doce"), preco_maximo = 50)

Arguments

preco_maximo

Value

Uma tibble com uma marmita para comprar!

Examples

sortear_marmita()

#> Refeição  sugerida: SEM-CARNE MOÍDA COM LEGUMES 
#> 
#> Ingredientes: Feijão carioca, arroz cateto e trigo integral 
#> 
#> Preço: R$ 20.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/sem-carne-moida-com-legumes

sortear_marmita(preco_maximo = 22)

#> Refeição  sugerida: TORTA DE MAÇÃ COM ESPECIARIAS 
#> 
#> Ingredientes: Açúcar Demerara, Aveia e Linhaça 
#> 
#> Preço: R$ 8.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/torta-de-mac-com-especiarias

sortear_marmita(tipo = "Sopa")

#> Refeição  sugerida: MINESTRONE 
#> 
#> Ingredientes: Espinafre, abobrinha e feijão branco 
#> 
#> Preço: R$ 16.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/minestrone

sortear_marmita(tipo = "Prato")

#> Refeição  sugerida: VEGETAIS ASIÁTICOS AO MOLHO DE COCO 
#> 
#> Ingredientes: Grão de bico, mix de quinoa, sementes e purê de cenoura 
#> 
#> Preço: R$ 21.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/vegetais-asiaticos-ao-molho-de-coco

sortear_marmita(tipo = "Doce")

#> Refeição  sugerida: BROWNIE DE CHOCOLATE E NOZES 
#> 
#> Ingredientes: Chocolate 100% cacau 
#> 
#> Preço: R$ 8.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/brownie-de-chocolate-e-nozes